use Haskeline for readline-style input
We aren't doing anything useful with this yet.
Next we ought to look at custom keybinds next for reevaluate and undo,
and figure out the completion story.
Completion could trigger tab to subprocess, which would hopefully write
out all completions to stdout since isatty is false, if not we'll
need to page through completions manually.
For completion display, printing utop-style (a fancy OCaml repl)
is probably a good start. Might need to trick Haskeline into
staying on the terminal row where it was.